requirements.txt
# CIMMS - Computerised Instrument Maintenance and Monitoring System
# Backend & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data & DB
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# MQTT
paho-mqtt>=2.0.0

# Machine Learning & Analysis
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Utils
python-multipart>=0.0.6
python-dateutil>=2.8.2
reportlab>=4.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
